+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [0.4.2]
+
+
 ## [0.4.1]
 
 - fixed prompt field answer handling to properly prepare for next prompt after answering question